requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyodbc==4.0.39
Werkzeug==2.3.7
python-dotenv==1.0.0
